feat: Move the car forward
Passed the next values for the x and y coordinates of the car to the
simulator. With points set 0.5 meters part and the car moving 50 times
per second, this will create a constant velocity of 25 m/sec.
Note that future improvements will be needed to correct the spike in the
car acceleration, as well.